integrating Issue24Tests in main class PaymentDetailsTest
